requirements.txt
numpy~=1.18.1
torch
matplotlib~=3.3.0
torchvision~=0.6.0
torchsummary
pandas
biopython
import-ipynb
scikit-learn